clean and refactor some code
