+++ android/src/main/kotlin/com/github/doandadr/petualanganhijaiyah/android/ml/AndroidPrediction.kt
@@ -20,10 +20,6 @@ class AndroidRecognition(private val model: Model) : Recognition {
         }
         inputBuffer.loadArray(inputArray)
 
-//        // Prepare the input
-//        val inputTensorImage = TensorImage(DataType.FLOAT32)
-//        inputTensorImage.load(inputBuffer)
-
         // Run the model
         val outputs = model.process(inputBuffer)
         val outputBuffer = outputs.outputFeature0AsTensorBuffer
