+++ core/src/main/kotlin/com/github/doandadr/petualanganhijaiyah/asset/MusicAsset.kt
@@ -5,11 +5,11 @@ import com.badlogic.gdx.audio.Music
 
 enum class MusicAsset(
     fileName: String,
-    val volume: Float = 0.5f,
+    val volume: Float = 0.6f,
     directory: String = "audio/music",
     val descriptor: AssetDescriptor<Music> = AssetDescriptor("$directory/$fileName", Music::class.java)
 ) {
-    DESERT("desert.mp3", 0.15f),
+    DESERT("desert.mp3", 0.15f), // TODO copyrighted
     DESERT_ALT("desert-2.mp3",0.4f),
     FIELD("forest.mp3"),
     HOME("home.mp3", 0.4f),
