+++ core/src/main/kotlin/com/github/doandadr/petualanganhijaiyah/data/LevelModel.kt
@@ -1,0 +1,14 @@
+package com.github.doandadr.petualanganhijaiyah.data
+
+data class LevelModel(
+    val number: Int,
+    val name: String,
+    val isTimed: Boolean,
+    val timerSeconds: Float,
+    val isHealthCounted: Boolean,
+    val maxHealth: Int,
+    val isScored: Boolean,
+    val stages: List<StageModel>,
+    val bgIndex: Int,
+    val musicIndex: Int,
+)
