+++ core/src/main/kotlin/com/github/doandadr/petualanganhijaiyah/data/StageModel.kt
@@ -1,0 +1,7 @@
+package com.github.doandadr.petualanganhijaiyah.data
+
+data class StageModel(
+    val type: StageType,
+    val choices: Int = 3,
+    val rounds: Int = 5,
+)
