+++ core/src/main/kotlin/com/github/doandadr/petualanganhijaiyah/data/StageType.kt
@@ -1,0 +1,10 @@
+package com.github.doandadr.petualanganhijaiyah.data
+
+enum class StageType(val title: String) {
+    MCQ("Tebak Huruf"),
+    MCQ_JOIN("Sambungan"),
+    MCQ_VOICE("Ucapan"),
+    DRAG_AND_DROP("Cetakan"),
+    MATCH_LINE("Tarik Garis"),
+    DRAWING("Menggambar"),
+}
