+++ core/src/main/kotlin/com/github/doandadr/petualanganhijaiyah/screen/PracticeScreen.kt
@@ -5,7 +5,6 @@ import com.badlogic.gdx.utils.Align
 import com.github.doandadr.petualanganhijaiyah.Main
 import com.github.doandadr.petualanganhijaiyah.asset.ImageButtons
 import com.github.doandadr.petualanganhijaiyah.asset.Labels
-import com.github.doandadr.petualanganhijaiyah.asset.MusicAsset
 import com.github.doandadr.petualanganhijaiyah.asset.TextureAsset
 import com.github.doandadr.petualanganhijaiyah.ui.animation.Animations
 import com.github.doandadr.petualanganhijaiyah.ui.values.PADDING_INNER_SCREEN
@@ -15,7 +14,10 @@ import ktx.actors.onChange
 import ktx.actors.onTouchDown
 import ktx.actors.plusAssign
 import ktx.log.logger
-import ktx.scene2d.*
+import ktx.scene2d.actors
+import ktx.scene2d.imageButton
+import ktx.scene2d.label
+import ktx.scene2d.table
 
 private val log = logger<PracticeScreen>()
 
@@ -28,10 +30,7 @@ class PracticeScreen(game: Main) : BaseScreen(game) {
     }
 
     private fun setupUI() {
-        val skin = Scene2DSkin.defaultSkin
         val bgPractice = assets[TextureAsset.STAGE.descriptor]
-
-        audioService.play(MusicAsset.FIELD)
 
         stage.actors {
             table {
@@ -61,12 +60,6 @@ class PracticeScreen(game: Main) : BaseScreen(game) {
                 practiceStage(assets, audioService, gameEventManager) {
                     //it.prefSize(STAGE_BOX_WIDTH, STAGE_BOX_HEIGHT)
                     it.grow()
-                }
-
-                row()
-                imageButton(ImageButtons.QUESTION.style) {
-                    it.expand().align(Align.bottomRight).padBottom(PADDING_INNER_SCREEN).padRight(PADDING_INNER_SCREEN)
-                    isVisible = false
                 }
             }
         }
